// src/app/services/model/projectExpenditureResponseModel.ts
/**
 * Metal Casting Project API
 * This API serves the frontend with data from databases and other external services.
 *
 * The version of the OpenAPI document: v1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ProjectExpenditureResponseModel { 
    id?: number;
    projectNumber?: string | null;
    focusArea?: string | null;
    item?: string | null;
    actualCost?: number;
    targetCost?: number;
}